refactor: unifying the getWeddings hook call into only one component
