+++ frontend/src/app/portal/dashboard/(components)/weddingsDisplay/guest-weddings.tsx
@@ -1,21 +1,12 @@
 'use client'
 
 import WeddingCard from '../weddingCard/wedding-card'
-import { useEffect } from 'react'
 import { useWeddingsStore } from '@/stores/weddings/weddings.provider'
 import { useShallow } from 'zustand/shallow'
 import { ClipLoader } from 'react-spinners'
 import { useGeneralStore } from '@/stores/general/general.provider'
-import useGetWeddings from '../../(hooks)/useGetWeddings'
 
 export default function WeddingsGuest() {
-  const getWeddings = useGetWeddings()
-
-  useEffect(() => {
-    getWeddings()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
   const { invitedWeddings, hasHydrated } = useWeddingsStore(
     useShallow((store) => ({
       invitedWeddings: store.invitedWeddings,
