+++ frontend/src/app/portal/dashboard/page.tsx
@@ -1,23 +1,5 @@
-import WeddingsOwn from './(components)/weddingsDisplay/own-weddings'
-import WeddingsGuest from './(components)/weddingsDisplay/guest-weddings'
+import DashboardView from './(components)/dashboard-view/dashboard-view'
 
 export default function Dashboard() {
-  return (
-    <>
-      <div className="grid lg:grid-cols-2 gap-5 lg:gap-20 items-center md:items-start w-full pb-5 md:px-20 xl:px-40">
-        <div className="flex flex-col gap-6 items-center min-h-[250px]">
-          <h1 className="font-bold font-poppins text-amber-800 text-2xl">
-            My Weddings:
-          </h1>
-          <WeddingsOwn />
-        </div>
-        <div className="flex flex-col gap-6 items-center min-h-[250px]">
-          <h1 className="font-bold font-poppins text-amber-800 text-2xl">
-            Weddings I'm a guest at:
-          </h1>
-          <WeddingsGuest />
-        </div>
-      </div>
-    </>
-  )
+  return <DashboardView />
 }
